RAG Pipeline with Reranking & Eval/requirements.txt
langchain>=0.0.200
openai>=1.0.0
chromadb>=0.4.0
faiss-cpu>=1.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
numpy>=1.24.0
scikit-learn>=1.3.0
pypdf>=4.2.0
python-dotenv>=1.0.1
rank-bm25>=0.2.2
sentence-transformers>=2.6.0
